+++ src/ft_strncmp.c
@@ -1,6 +1,8 @@
-int ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    unsigned int i;
+    size_t i;
 
     i = 0;
     while (s1[i] && s2[i] && s1[i] == s2[i] && i < n - 1)
